Created a new version of the IDE with a better editor. Modified the CodeRunner class such that the code can run in another process (better for program stability).

+++ helpers/coderunner.py
@@ -1,5 +1,6 @@
 import sys
-
+import time
+import traceback
 from pyview.lib.patterns import KillableThread,Reloadable,StopThread,Subject
 from threading import RLock
 
@@ -13,6 +14,7 @@ class CodeThread (KillableThread):
     self._failed = False
     self._callback = callback
     self.code = code
+    self._stop = False
     
   def isRunning(self):
     return self.isAlive()
@@ -23,7 +25,15 @@ class CodeThread (KillableThread):
   def exceptionInfo(self):
     if self.failed() == False:
       return None
-    return (self._exception_type,self._exception_value,self._traceback)
+    return (self._exception_type,self._exception_value)
+
+  def tracebackInfo(self):
+    if self.failed() == False:
+        return None
+    return self._traceback
+    
+  def stop(self):
+    self._stop = True
     
   def run(self):
     try:
@@ -32,11 +42,11 @@ class CodeThread (KillableThread):
     except StopThread:
       pass
     except:
+      self._failed = True
       exc_type, exc_value, exc_traceback = sys.exc_info()
       self._exception_type = exc_type
       self._exception_value = exc_value
       self._traceback = exc_traceback
-      self._failed = True
       raise
     finally:
       if not self._callback == None:
@@ -54,19 +64,34 @@ class CodeRunner(Reloadable,Subject):
   def __init__(self,gv = dict(),lv = dict()):
     Reloadable.__init__(self)
     Subject.__init__(self)
+    self._threadID = 0
     self.clear(gv = gv,lv = lv)
     
   def clear(self,gv = dict(),lv = dict()):
     self._gv = gv
     self._lv = dict()
-    self._threads = dict()
-    self._exceptions = []
+    self._threads = {}
+    self._exceptions = {}
+    self._tracebacks = {}
     
   def clearExceptions(self):
-    self._exceptions = []
+    self._exceptions = {}
     
-  def exceptions(self):
-    return self._exceptions
+  def getException(self,identifier):
+    if identifier in self._exceptions:
+      return self._exceptions[identifier]
+    return None
+
+  def getTraceback(self,identifier):
+    if identifier in self._tracebacks:
+      return traceback.extract_tb(self._tracebacks[identifier])
+    return None
+
+  def formatException(self,identifier):
+    exc = self.exception(identifier)
+    tb = self.traceback(identifier)
+    return traceback.format_exception(exc[0],exc[1],tb)
+    
 
   def gv(self):
     return self._gv
@@ -75,8 +100,8 @@ class CodeRunner(Reloadable,Subject):
     lock = RLock()
     lock.acquire()
     if thread.failed():
-      self._exceptions.append(thread.exceptionInfo())
-      self.notify("exceptions",self.exceptions())
+      self._exceptions[thread._id] = thread.exceptionInfo()
+      self._tracebacks[thread._id] = thread.tracebackInfo()
     lock.release()
     
   def lv(self):
@@ -86,7 +111,12 @@ class CodeRunner(Reloadable,Subject):
     if identifier in self._threads:
       return self._threads[identifier].failed()
       
-  def isExecutingCode(self,identifier):
+  def isExecutingCode(self,identifier = None):
+    if identifier == None:
+      for thread in self._threads.values():
+        if thread.isRunning():
+          return True
+      return False
     if not identifier in self._threads:
       return False
     if self._threads[identifier] == None:
@@ -102,7 +132,7 @@ class CodeRunner(Reloadable,Subject):
     
   def executeCode(self,code,identifier,filename = None, lv = None,gv = None):
     if self.isExecutingCode(identifier):
-      return False
+      return -1
     if lv == None:
       if not identifier in self._lv:
         self._lv[identifier] = dict()
@@ -120,10 +150,12 @@ class CodeRunner(Reloadable,Subject):
     lv["gv"] = gvClass
     
     ct = CodeThread(code,source = filename,lv = lv,gv = lv,callback = self.threadCallback)
+    ct._id = self._threadID
+    self._threadID+=1
     self._threads[identifier] = ct
     ct.setDaemon(True)
     ct.start()
-    return True
+    return ct._id
     
 from multiprocessing import *
 from pyview.helpers.coderunner import *
@@ -141,12 +173,20 @@ class CodeProcess(Process):
     def write(self,output):
       self._queue.put(output)
       
+    def read(self):
+      if not self._queue.empty():
+        return self._queue.get(False)
+      else:
+        return ""
+      
   def __init__(self):
     Process.__init__(self)
+    self.daemon = True
     self._commandQueue = Queue()
     self._responseQueue = Queue()
     self._stdoutQueue = Queue()
     self._stderrQueue = Queue()
+    self._stdinQueue = Queue()
     self._codeRunner = CodeRunner()
     
   def commandQueue(self):
@@ -155,6 +195,9 @@ class CodeProcess(Process):
   def responseQueue(self):
     return self._responseQueue
     
+  def stdinQueue(self):
+    return self._stdinQueue
+    
   def stdoutQueue(self):
     return self._stdoutQueue
     
@@ -162,40 +205,79 @@ class CodeProcess(Process):
     return self._stderrQueue
     
   def run(self):
+    print "New code process up and running..."
     sys.stderr = self.StreamProxy(self._stderrQueue)
     sys.stdout = self.StreamProxy(self._stdoutQueue)
+    sys.stdin = self.StreamProxy(self._stdinQueue)
     while True:
-      while not self.commandQueue().empty():
-        (command,args,kwargs) = self.commandQueue().get()
-        if command == "stop":
-          exit(0)
-        if hasattr(self._codeRunner,command):
-          f = getattr(self._codeRunner,command)
-          r = f(*args,**kwargs)
-          self.responseQueue().put(r)
-
-    
+      time.sleep(0.1)
+      try:
+        while not self.commandQueue().empty():
+          (command,args,kwargs) = self.commandQueue().get(False)
+          if command == "stop":
+            exit(0)
+          if hasattr(self._codeRunner,command):
+            try:
+              f = getattr(self._codeRunner,command)
+              r = f(*args,**kwargs)
+              self.responseQueue().put(r,False)
+            except Exception as e:
+              self.responseQueue().put(e,False)
+      except KeyboardInterrupt:
+        print "Interrupt, exiting..."
+      
 class MultiProcessCodeRunner():
   
   def __init__(self,gv = dict(),lv = dict()):
     self._codeProcess = CodeProcess()
     self._codeProcess.start()
+
+  def __del__(self):
+    self.terminate()
     
   def dispatch(self,command,*args,**kwargs):
     message = (command,args,kwargs)
+    if not self._codeProcess.is_alive():
+      self.restart()
     self._codeProcess.commandQueue().put(message)
     try:
-      response = self._codeProcess.responseQueue().get(timeout = 2)
+      response = self._codeProcess.responseQueue().get()
     except:
       response = None
     return response
+
+  def stdin(self,input):
+    self._codeProcess.stdinQueue().put(input)
+
+  def hasStdout(self):
+    return not self._codeProcess.stdoutQueue().empty()
+
+  def stdout(self):
+    string = ""
+    while not self._codeProcess.stdoutQueue().empty():
+      string+=self._codeProcess.stdoutQueue().get(False)
+    return string
+
+  def hasStderr(self):
+    return not self._codeProcess.stderrQueue().empty()
+
+  def stderr(self):
+    string = ""
+    while not self._codeProcess.stderrQueue().empty():
+      string+=self._codeProcess.stderrQueue().get(False)
+    return string
     
   def codeProcess(self):
     return self._codeProcess
     
   def stop(self):
-    
     self._codeProcess.commandQueue().put(("stop",[],{}))
+    
+  def start(self):
+    if self._codeProcess.is_alive():
+      self.restart()
+    else:
+      self._codeProcess.start()
     
   def restart(self):
     print "Restarting code runner..."
@@ -204,7 +286,8 @@ class MultiProcessCodeRunner():
     self._codeProcess.start()
     
   def terminate(self):
-    self._codeProcess.terminate()
+    if self._codeProcess.is_alive():
+      self._codeProcess.terminate()
     
   def __getattr__(self,attr):
     return lambda *args,**kwargs: self.dispatch(attr,*args,**kwargs)
